+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## 1.0.0 under development
 
 - Initial release.
+- New #299: Add `hasStatusSupport()` method to `AdapterInterface` (WarLikeLaux)